moved TestResultStatus
