+++ save-agent/save-cloud-agent-api/src/commonMain/kotlin/com/saveourtool/save/agent/TestExecutionResult.kt
@@ -1,5 +1,6 @@
 package com.saveourtool.save.agent
 
+import com.saveourtool.save.domain.TestResultStatus
 import kotlinx.serialization.Serializable
 
 /**
