+++ save-cloud-common/src/commonMain/kotlin/com/saveourtool/save/agent/TestExecutionDto.kt
@@ -1,5 +1,6 @@
 package com.saveourtool.save.agent
 
+import com.saveourtool.save.domain.TestResultStatus
 import com.saveourtool.save.entities.DtoWithId
 import com.saveourtool.save.test.analysis.metrics.NoDataAvailable
 import com.saveourtool.save.test.analysis.metrics.TestMetrics
