+++ save-cloud-common/src/commonMain/kotlin/com/saveourtool/save/agent/TestSuiteExecutionStatisticDto.kt
@@ -1,5 +1,6 @@
 package com.saveourtool.save.agent
 
+import com.saveourtool.save.domain.TestResultStatus
 import kotlinx.serialization.Serializable
 
 /**
